+++ Frontend/src/lib/icons/index.ts
@@ -2,6 +2,10 @@ export {default as SuccessIcon} from './Success.svelte';
 export {default as AlertIcon} from './Alert.svelte';
 export {default as InfoIcon} from './Info.svelte';
 
+export {default as VerificationIcon} from './Verification.svelte';
+export {default as PasswordIcon} from './Password.svelte';
+export {default as UserIcon} from './User.svelte';
+
 export {default as CheckmarkIcon} from './Checkmark.svelte';
 export {default as QuestionIcon} from './Question.svelte';
 export {default as ChevronIcon} from './Chevron.svelte';