+++ Frontend/src/lib/components/index.ts
@@ -5,6 +5,9 @@ export {default as Footer} from './layout/Footer.svelte';
 
 export {default as Card} from './main/Card.svelte';
 export {default as Input} from './main/Input.svelte';
+export {default as Loader} from './main/Loader.svelte';
+export {default as Button} from './main/Button.svelte';
+
 export {default as Steps} from './main/Steps.svelte';
 export {default as StepsItem} from './main/StepsItem.svelte';
 export {default as SelectMenu} from './main/SelectMenu.svelte';